Add abstract storage with incremental ids
